backend/requirements.txt
alembic==1.14.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.8
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.23
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3
